backend/requirements.txt
Django==5.0.2
djangorestframework==3.14.0
django-cors-headers==4.3.1
sqlparse==0.4.4
tzdata==2024.1
django-filter==23.5
python-dotenv==1.0.1
Pillow==10.2.0
psycopg2-binary==2.9.9
gunicorn==21.2.0
whitenoise==6.6.0 